requirements.txt
# Web framework
fastapi==0.136.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.49
psycopg2-binary==2.9.10
alembic==1.16.1

# Auth & security
pyjwt==2.12.1
pwdlib[argon2]==0.3.0
python-dotenv==1.1.0

# Pydantic
pydantic==2.11.4
pydantic-settings==2.9.1
email-validator==2.3.0

# Google integrations
google-auth==2.40.1
google-genai>=0.8.0

# QR code
qrcode[pil]==8.2

# Rate limiting
slowapi==0.1.9

# Redis (optional — shared rate-limit + cache across workers)
redis>=5.0.0

# Dev / CI
ruff>=0.9.0